+++ src/main-tvc.cpp
@@ -272,7 +272,7 @@ void loop() {
 
                 if (millis() % 1000 < 40) {
                     rocket.printMessage("Launching in T -", false);
-                    rocket.printMessage((int64_t)launchTimer, false);
+                    rocket.printMessage(launchTimer, false);
                     rocket.printMessage("seconds! Press any key + ENTER to abort the launch.");
 
                     launchTimer--;
@@ -349,7 +349,7 @@ void loop() {
         if (millis() - lastStepMs > 100) {
             onGroundSteps++;
             rocket.printMessage("On ground for ", false);
-            rocket.printMessage((int64_t)onGroundSteps, false);
+            rocket.printMessage(onGroundSteps, false);
             rocket.printMessage(" /20 steps (", false);
             rocket.printMessage((float)onGroundSteps * 0.1f, false);
             rocket.printMessage(") sec.");
